requirements.txt
fastapi==0.89.1
mongoengine==0.24.2
odmantic==0.9.2
python-dotenv==0.21.0
uvicorn==0.20.0
